Refator toc
